+++ point.h
@@ -20,15 +20,15 @@ namespace mathTools
 /*-----------------------------------------*/
 
 			//Contains the spacial information for x,y,z cordinates.
-			float * x;
-			float * y;
-			float * z;
+			double * x;
+			double * y;
+			double * z;
 			//Contains the velocity information for x,y,z cordinates.
-			float * vx;
-			float * vy;
-			float * vz;
+			double * vx;
+			double * vy;
+			double * vz;
 			//Contains the radius of each particle.
-			float r;
+			double r;
 
 			//Contains the system information.
 			int boxSize;
@@ -37,7 +37,7 @@ namespace mathTools
 			int seed;
 
 			//Initial system temperature
-			float t_init;
+			double t_init;
 
 			//Base initialization
 			void init();
@@ -52,7 +52,7 @@ namespace mathTools
 			int arrSize;
 
 			//Constructor/Destructor
-			points(int nParticles, float radius, float t_initial);
+			points(int nParticles, double radius, double t_initial);
 			points( const points &obj );
 			~points();
 
@@ -61,33 +61,33 @@ namespace mathTools
 /*-----------------------------------------*/
 
 			//Getters for each spacial cordinate.
-			float getX(int index) { return *(x+index); }
-			float getY(int index) { return *(y+index); }
-			float getZ(int index) { return *(z+index); }
-			float getR() { return r; }
-			float getBoxSize() { return boxSize; }
+			double getX(int index) { return *(x+index); }
+			double getY(int index) { return *(y+index); }
+			double getZ(int index) { return *(z+index); }
+			double getR() { return r; }
+			double getBoxSize() { return boxSize; }
 
 			//Getters for each velocity cordinate.
-			float getVX(int index) { return *(vx+index); }
-			float getVY(int index) { return *(vy+index); }
-			float getVZ(int index) { return *(vz+index); }
+			double getVX(int index) { return *(vx+index); }
+			double getVY(int index) { return *(vy+index); }
+			double getVZ(int index) { return *(vz+index); }
 
 /*-----------------------------------------*/
 /*--------------SYSTEM SETTERS-------------*/
 /*-----------------------------------------*/
 
 			//Standard setters for position
-			void setX( int index, float val ) { *(x+index) = utilities::safeMod(val,boxSize); }
-			void setY( int index, float val ) { *(y+index) = utilities::safeMod(val,boxSize); }
-			void setZ( int index, float val ) { *(z+index) = utilities::safeMod(val,boxSize); }
-			void setR( float val ) { r = val; }
-			void setAllPos( int index, float xVal, float yVal, float zVal);
+			void setX( int index, double val ) { *(x+index) = utilities::safeMod(val,boxSize); }
+			void setY( int index, double val ) { *(y+index) = utilities::safeMod(val,boxSize); }
+			void setZ( int index, double val ) { *(z+index) = utilities::safeMod(val,boxSize); }
+			void setR( double val ) { r = val; }
+			void setAllPos( int index, double xVal, double yVal, double zVal);
 
 			//Setters for velocity
-			void setVX( int index, float val ) { *(vx+index) = val; }
-			void setVY( int index, float val ) { *(vy+index) = val; }
-			void setVZ( int index, float val ) { *(vz+index) = val; }
-			void setAllVel( int index, float vxVal, float vyVal, float vzVal);
+			void setVX( int index, double val ) { *(vx+index) = val; }
+			void setVY( int index, double val ) { *(vy+index) = val; }
+			void setVZ( int index, double val ) { *(vz+index) = val; }
+			void setAllVel( int index, double vxVal, double vyVal, double vzVal);
 
 /*-----------------------------------------*/
 /*---------------SYSTEM DEBUG--------------*/
@@ -102,8 +102,8 @@ namespace mathTools
 /*-----------------------------------------*/
 
 			//Creates an initial distribution of the particles.
-			void init(float concentration);
-			void init(float concentration, int seedling);
+			void init(double concentration);
+			void init(double concentration, int seedling);
 			void initCheck(std::mt19937* gen, std::uniform_real_distribution<double>* distribution);
 			void maxwellVelocityInit(std::mt19937* gen, std::uniform_real_distribution<double>* distribution);
