+++ verlet.cpp
@@ -3,7 +3,7 @@
 namespace integrators
 {
 	//Sets time step and zeros 'systemTime'.
-	verlet::verlet(float dt)
+	verlet::verlet(double dt)
 	{
 		timeStep = dt;
 		systemTime = 0.0;
@@ -17,20 +17,20 @@ namespace integrators
 	}
 
 	//Calculates the next position and velocity
-	int verlet::nextSystem(float index, mathTools::points* pts, physics::forces* f)
+	int verlet::nextSystem(double index, mathTools::points* pts, physics::forces* f)
 	{
 		//A(t)
-		float acc[3] = {0.0,0.0,0.0};
+		double acc[3] = {0.0,0.0,0.0};
 		//A(t+dt)
-		float accNext[3] = {0.0,0.0,0.0};
+		double accNext[3] = {0.0,0.0,0.0};
 
 		//Gets A(t)
 		f->getAcceleration(index,systemTime,pts,acc);
 
 		//Gets the new position.
-		float xNew = posAlgorithm(pts->getX(index),pts->getVX(index),acc[0],systemTime);
-		float yNew = posAlgorithm(pts->getY(index),pts->getVY(index),acc[1],systemTime);
-		float zNew = posAlgorithm(pts->getZ(index),pts->getVZ(index),acc[2],systemTime);
+		double xNew = posAlgorithm(pts->getX(index),pts->getVX(index),acc[0],systemTime);
+		double yNew = posAlgorithm(pts->getY(index),pts->getVY(index),acc[1],systemTime);
+		double zNew = posAlgorithm(pts->getZ(index),pts->getVZ(index),acc[2],systemTime);
 
 		pts->setX(index, xNew);
 		pts->setY(index, yNew);
@@ -49,14 +49,14 @@ namespace integrators
 
 	//The velocity verlet Algorithms to obtain the next position term.
 	//Performs operation on one spacial cordinate.
-	float verlet::posAlgorithm(float pos, float vel, float acc, float t)
+	double verlet::posAlgorithm(double pos, double vel, double acc, double t)
 	{
 		return pos+(vel*timeStep)+0.5*timeStep*timeStep*acc;
 	}
 
 	//The velocity verlet Algorithms to obtain the next velocity term term.
 	//Performs operation on one velocity cordinate.
-	float verlet::velAlgorithm(float pos, float vel, float acc, float accnext, float t)
+	double verlet::velAlgorithm(double pos, double vel, double acc, double accnext, double t)
 	{
 		return vel+0.5*timeStep*(acc+accnext);
 	}
